tests/test_cloudfront_integration.py
"""
Tests for CloudFront integration functionality.
"""

import pytest
from aws_cdk import App, Stack
from aws_cdk.assertions import Template
from stacks.user_interface.cloudfront import CloudFrontVpcOriginConstruct
from stacks.common.constants import UI_ACCESS_MODE_PUBLIC, UI_ACCESS_MODE_PRIVATE


class TestCloudFrontIntegration:
    """Test CloudFront VPC origins integration."""

    def test_cloudfront_construct_creation(self):
        """Test that CloudFront construct can be created."""
        app = App()
        stack = Stack(app, "TestStack")
        
        # This test would require a mock ALB, but validates the basic structure
        # In a full implementation, we'd create a mock ALB and test the construct
        assert True  # Placeholder for actual construct creation test

    def test_access_mode_validation(self):
        """Test UIAccessMode validation logic."""
        from stacks.user_interface.stack import WebAppStack
        
        valid_modes = [UI_ACCESS_MODE_PUBLIC, UI_ACCESS_MODE_PRIVATE]
        
        # Test valid modes
        for mode in valid_modes:
            # This should not raise an exception
            assert mode in ["public", "private"]
        
        # Test invalid mode would raise ValueError
        invalid_mode = "invalid"
        assert invalid_mode not in valid_modes

    def test_waf_conditional_creation(self):
        """Test that WAF is created conditionally based on access mode."""
        # Test that WAF placement logic works correctly
        # In public mode: WAF on CloudFront
        # In private mode: WAF on ALB
        assert UI_ACCESS_MODE_PUBLIC == "public"
        assert UI_ACCESS_MODE_PRIVATE == "private"

    def test_alb_security_group_configuration(self):
        """Test ALB security group configuration for different access modes."""
        # Test that security groups are configured correctly
        # Public mode: CloudFront prefix lists only
        # Private mode: Customer prefix lists
        assert True  # Placeholder for security group configuration test
