requirements.txt
aws-cdk-lib==2.208.0
constructs
pyyaml
cdk-nag
boto3